TRex Bomb Ai added plus more
In this version all of the Ai dinos only go after the one
player-controlled dino in 'single player' mode, or in 'multi-player'
mode they will go after ANY of the player-controlled dinos equally.
Version Notes/Changes:
1. Ai TRex Bomb, Napalm, function added to AIControl.cs
2. AIHesp.cs logic all have been moved and implemented in the
AIControl.cs script
3. Removed AIHesp.cs script from Hesp.prefab
4. Deleted AIHesp.cs script from project
5. Removed lines of code from RaceStart.cs that activated the AIHesp.cs
script for Hesp Ai dinos
6. Changed one joystick button in the InputManager.asset for better
controls on my joystick I'm using.
Another note: I need to figure out why, when a Troodon Ai is in the
race, it creates a couple to a few hundred errors that say "must be
non-kinematic object". My first instinct is that the Troodon Ai Melee
function is picking up all objects surrounding him whether they are
dinos or pieces of the track etc.. I believe it should be an easy fix.